+++ R/generics.R
@@ -33,14 +33,6 @@ setGeneric(
 )
 
 
-# Methods and documentations found in methods-spatShift.R
-setGeneric("spatShift", function(x, ...) standardGeneric("spatShift"))
-setGeneric("affine", function(x, y, ...) standardGeneric("affine"))
-setGeneric("shear", function(x, ...) standardGeneric("shear"))
-
-# Methods and documentations found in methods-overlaps.R
-setGeneric("overlaps", function(x, ...) standardGeneric("overlaps"))
-
 
 # Object creation ####
 setGeneric(
@@ -86,6 +78,16 @@ setGeneric(
     "overlapToMatrix",
     function(x, ...) standardGeneric("overlapToMatrix")
 )
+
+# Methods and documentations found in methods-spatShift.R
+setGeneric("spatShift", function(x, ...) standardGeneric("spatShift"))
+setGeneric("affine", function(x, y, ...) standardGeneric("affine"))
+setGeneric("shear", function(x, ...) standardGeneric("shear"))
+setGeneric("XY", function(x, ...) standardGeneric("XY"))
+setGeneric("XY<-", function(x, ..., value) standardGeneric("XY<-"))
+
+# Methods and documentations found in methods-overlaps.R
+setGeneric("overlaps", function(x, ...) standardGeneric("overlaps"))
 
 
 # Giotto subnesting ####